+++ src/config.ts
@@ -306,7 +306,6 @@ if (isDebugMode || isQuickMode) {
     Config.hashtagTootRetrievalDelaySeconds = 2;
     Config.incrementalLoadDelayMS = 100;
     Config.lookbackForUpdatesMinutes = 15;
-    Config.maxCachedTimelineToots = 700;
     Config.maxRecordsForFeatureScoring = 480;
     Config.numDesiredTimelineToots = 500;
     Config.numParticipatedTagsToFetchTootsFor = 10;