+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
-* Fix bug where getParticipatedHashtag() toots wasn't returning a promise
+* Add `moreOldToots` parameter to `triggerFeedUpdate()`
+* Fix bug where `getParticipatedHashtag()` toots wasn't returning a promise
 * Up `Config.maxCachedTimelineToots` to 2,500
 * Remove unused `Config.reloadFeaturesEveryNthOpen` param